+++ src/samples/Sandbox/Game/SpaceShooter.xaml
@@ -1,10 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<game:MauiGame
+<game1:MauiGame
     x:Class="SpaceShooter.Game.SpaceShooter"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:draw="http://schemas.appomobi.com/drawnUi/2023/draw"
     xmlns:game="clr-namespace:SpaceShooter.Game"
+    xmlns:game1="clr-namespace:DrawnUi.Maui.Game;assembly=DrawnUi.Maui.Game"
     x:DataType="game:SpaceShooter"
     BackgroundColor="#000011"
     HorizontalOptions="Fill"
@@ -373,4 +374,4 @@
 
     </draw:SkiaLayout>
 
-</game:MauiGame>
+</game1:MauiGame>
